+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog]
 
 ## [Unreleased]
 
+- A service operator can put an individual service into maintenance mode from
+  the admin interface
+
 ## [Release 030] - 2019-11-12
 
 - Add policy specific "Reply-to" address for claim emails